Split the test cases into separate files
